% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomes.R
\name{chromnames}
\alias{chromnames}
\title{Get chromosome names}
\usage{
chromnames(clean_names = FALSE, genome = "pf3d7", ...)
}
\arguments{
\item{clean_names}{logical; if \code{TRUE}, scrub names like 'Pf3D7_01_v3','Pf3D7_02_v3' to '1','2'}

\item{genome}{with what genome are we working?}

\item{...}{ignored}
}
\value{
the chromosome names of the desired genome, in kartoype order
}
\description{
Get chromosome names
}
