% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomes.R
\name{chromsizes}
\alias{chromsizes}
\alias{chromsizes_3d7}
\alias{chromsizes_sal1}
\alias{chromsizes_p01}
\alias{chromsizes_cM}
\title{Get named vector of chromosome sizes in specified genome assembly}
\usage{
chromsizes(genome = "pf3d7", clean_names = FALSE, as_seqinfo = FALSE,
  ...)

chromsizes_3d7(...)

chromsizes_sal1(...)

chromsizes_p01(...)

chromsizes_cM(..., clean_names = FALSE, genome = "pf3d7")
}
\arguments{
\item{genome}{identifier for a genome assembly}

\item{clean_names}{logical; if \code{TRUE}, scrub chromosome names like 'Pf3D7_01_v3','Pf3D7_02_v3' to '1','2'}

\item{as_seqinfo}{logical; if \code{TRUE}, return a \code{GenomeInfoDb::Seqinfo} object instead of a named vector}

\item{...}{ignored}
}
\value{
a named vector of chromosome lengths in base pairs or (in case of \code{chromsizes_cM()}) in centimorgans
}
\description{
Get named vector of chromosome sizes in specified genome assembly
}
