% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomes.R
\name{.match_genome}
\alias{.match_genome}
\title{Guess a specific genome assembly from given character string.}
\usage{
.match_genome(g)
}
\description{
Guess a specific genome assembly from given character string.
}
