% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{encode_country}
\alias{encode_country}
\alias{encode_continent}
\title{Convert country names to standard codes}
\usage{
encode_country(x, code = c("iso2c", "iso3c", "iso.name", "un.name",
  "country.name", "fips.name", "genc.name", "continent"), ...)

encode_continent(x, ...)
}
\arguments{
\item{x}{vector of country identifiers to convert}

\item{code}{scheme for encoding countries in the output; default is ISO 2-letter code, but
see \code{?countrycode::codelist} for other choices}

\item{...}{ignored}
}
\value{
character vector of country identifiers in the requested encoding; missing or unmatched values will be \code{NA}.
  In the case of \code{encode_continent()}, a character vector of continents as assigned by the World Bank.
}
\description{
Convert country names to standard codes
}
\details{
All the heavy lifting is done by the \code{countrycode} package. This function will attempt to find
  matches for the input identifiers against a heirarchy of country encodings, starting with ISO 2-letter codes --
  so it should be idempotent with respect to already-encoded input.
}
\seealso{
\link[countrycode]{codelist}, \link[countrycode]{countrycode}
}
