% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomes.R
\name{factor_chrom}
\alias{factor_chrom}
\title{Sanitize chromosome names}
\usage{
factor_chrom(x, genome = "Pf3D7", ...)
}
\description{
Convert a vector of chromosome names in some sensible format to "nice" names in karyotype order.
}
