% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recomb_map_pf3d7}
\alias{recomb_map_pf3d7}
\title{Genetic map for P. falciparum genome}
\format{A \code{tibble} (inherits from data frame) with 4 columns:
\describe{
  \item{chr}{chromosome names, as ordered factor (\code{pf3d7} assembly)}
  \item{marker}{dummy marker name, created by concatenating chromosme and position of mid-point of crossover uncertainty interval}
  \item{cM}{genetic position in centimorgans}
  \item{pos}{physical position in base pairs on \code{pf3d7} assembly, 1-based}
}}
\source{
Miles A et al. (2016) \emph{Genome Research} 26: 1288-1299. \url{https://doi-org.libproxy.lib.unc.edu/10.1101/gr.203711.115}
}
\usage{
recomb_map_pf3d7
}
\description{
A genetic map (== recombination map) derived from crossovers reported in Miles et al. (2016), using identity
  map function (ie. no unobserved double-crossovers). Format follows the \code{*.map} format of PLINK and other
  programs for pedigree and linkage analysis.
}
\keyword{datasets}
