% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions_pf3d7}
\alias{regions_pf3d7}
\title{Segmentation of P. falciparum genome into "core" and hypervariable regions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 110 rows and 4 columns.}
\source{
Miles A et al. (2016) \emph{Genome Research} 26: 1288-1299. \url{https://doi-org.libproxy.lib.unc.edu/10.1101/gr.203711.115}
}
\usage{
regions_pf3d7
}
\description{
Segmentation of P. falciparum genome into "core" and hypervariable regions
}
\keyword{datasets}
