% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions_pvsal1}
\alias{regions_pvsal1}
\title{Segmentation of P. vivax genome into "core" and hypervariable regions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 47 rows and 4 columns.}
\source{
Pearson RD et al. (2016) \emph{Nature Genetics} 48: 959-964. \url{https://doi.org/10.1038/ng.3599}
}
\usage{
regions_pvsal1
}
\description{
Segmentation of P. vivax genome into "core" and hypervariable regions
}
\keyword{datasets}
