% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomes.R
\name{same_genome}
\alias{same_genome}
\title{Check if chromosome vectors are from same genome}
\usage{
same_genome(...)
}
\arguments{
\item{...}{any number of vectors encoding chromosomes; need not have same length}
}
\value{
logical scalar indicating whether the input vectors have all been encoded as chromosomes in same genome assembly
}
\description{
Check if chromosome vectors are from same genome
}
