% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scale_x_genome}
\alias{scale_x_genome}
\title{Sensible continuous horizontal axis for genomic coordinates}
\usage{
scale_x_genome(..., scale = 1000)
}
\description{
Sensible continuous horizontal axis for genomic coordinates
}
