% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomes.R
\name{chromsizes}
\alias{chromsizes}
\title{Get named vector of chromosome sizes in specified genome assembly}
\usage{
chromsizes(genome = "pf3d7", clean_names = FALSE, as_seqinfo = FALSE,
  ...)
}
\arguments{
\item{genome}{identifier for a genome assembly}

\item{clean_names}{logical; if \code{TRUE}, scrub chromosome names like 'Pf3D7_01_v3','Pf3D7_02_v3' to '1','2'}

\item{as_seqinfo}{logical; if \code{TRUE}, return a \code{GenomeInfoDb::Seqinfo} object instead of a named vector}

\item{...}{ignored}
}
\value{
a named vector of chromosome lengths in base pairs
}
\description{
Get named vector of chromosome sizes in specified genome assembly
}
