% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomes.R
\name{chromsizes}
\alias{chromsizes}
\title{Get named vector of chromosome sizes in specified genome assembly}
\usage{
chromsizes(genome = "Pf3D7", clean_names = FALSE, ...)
}
\description{
Get named vector of chromosome sizes in specified genome assembly
}
