% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosomes.R
\name{is_nuclear}
\alias{is_nuclear}
\title{Check if chromosome names are in nuclear genome}
\usage{
is_nuclear(x, species = c("pf", "pv"), ...)
}
\description{
Check if chromosome names are in nuclear genome
}
