% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scale_x_genome}
\alias{scale_x_genome}
\title{Sensible continuous horizontal axis for genomic coordinates}
\usage{
scale_x_genome(..., scale = 1000)
}
\arguments{
\item{...}{passed through to \code{ggplot2::scale_x_continuous()}}

\item{scale}{integer; label scale in units of this many base pairs (1e3 = kb, 1e6 = Mb, etc)}
}
\value{
a \code{ggplot2} scale object
}
\description{
Sensible continuous horizontal axis for genomic coordinates
}
